Created AuthGoogleService and moved some functions
