+++ prediction-polls/backend/src/services/AuthorizationService.js
@@ -1,112 +1,54 @@
 require('dotenv').config();
 const jwt = require("jsonwebtoken");
-const bcrypt = require('bcrypt')
-const qs = require("querystring");
-const axios = require("axios")
-
 const db = require("../repositories/AuthorizationDB.js");
-const { checkCredentials,addUser } = require('./AuthenticationService.js');
+
+const { checkCredentials, addUser } = require('./AuthenticationService.js');
+
+const bcrypt = require('bcrypt')
 
 
+function getGoogleOAuthURL() {
+  const rootUrl = "https://accounts.google.com/o/oauth2/v2/auth";
 
-function homePage(req, res){
-    res.json({"username":req.user.name,"key":"very secret"});
-  }
-
-function signup(req, res){
-  // Authorize User  
-  const { username, password, email, birthday } = req.body;
-  addUser( username, password, email, birthday );
-  
-}
-
-async function googleLogIn(req,res){
-  const code = req.query.code;
-
-  try {
-    // get the id and access token with the code
-    const { id_token, access_token } = await getGoogleOAuthTokens({ code });
-    console.log({ id_token, access_token });
-
-    // get user with tokens
-    const googleUser = await getGoogleUser({ id_token, access_token });
-
-    console.log({ googleUser });
-
-    if (!googleUser.verified_email) {
-      return res.status(403).send("Google account is not verified");
-    }
-
-    const user = {name : googleUser.name};
-    const accesToken = generateAccessToken(user);
-    const refreshToken = generateRefreshToken(user);
-    db.addRefreshToken(refreshToken);
-    res.json({accessToken: accesToken, refreshToken: refreshToken})
-
-  } catch (error) {
-    console.log(error, "Failed to authorize Google user");
-    return res.redirect("http://localhost:3000/login");
-  }
-}
-
-async function getGoogleOAuthTokens(code){
-  const url = "https://oauth2.googleapis.com/token";
-
-  const values = {
-    code,
-    client_id: process.env.googleClientId,
-    client_secret: process.env.googleClientSecret,
-    redirect_uri: process.env.googleOauthRedirectUrl,
-    grant_type: "authorization_code",
+  const options = {
+    redirect_uri: process.env.PUBLIC_GOOGLE_OAUTH_REDIRECT_URL,
+    client_id: process.env.PUBLIC_GOOGLE_CLIENT_ID,
+    access_type: "offline",
+    response_type: "code",
+    prompt: "consent",
+    scope: [
+      "https://www.googleapis.com/auth/userinfo.profile",
+      "https://www.googleapis.com/auth/userinfo.email",
+    ].join(" "),
   };
 
-  //Making the request
-  try {
-    const res = await axios.post<GoogleTokensResult>(
-      url,
-      qs.stringify(values),
-      {
-        headers: {
-          "Content-Type": "application/x-www-form-urlencoded",
-        },
-      }
-    );
-    return res.data;
-  } catch (error) {
-    console.error(error.response.data.error);
-    log.error(error, "Failed to fetch Google Oauth Tokens");
-    throw new Error(error.message);
-  }
+  const qs = new URLSearchParams(options);
 
+  return rootUrl + "?" + qs.toString();
 }
 
-async function getGoogleUser({
-  id_token,
-  access_token,
-}) {
-  try {
-    const res = await axios.get(
-      `https://www.googleapis.com/oauth2/v1/userinfo?alt=json&access_token=${access_token}`,
-      {
-        headers: {
-          Authorization: `Bearer ${id_token}`,
-        },
-      }
-    );
-    return res.data;
-  } catch (error) {
-    log.error(error, "Error fetching Google user");
-    throw new Error(error.message);
+function homePage(req, res){
+    const link = getGoogleOAuthURL()
+    res.status(200).json({"username":req.user.name,"key":link});
   }
-}
+
+async function signup(req, res){
+  console.log("Request Recieved");
+  const { username, password, email, birthday } = req.body;
+  const { success,error} = await addUser( username, password, email, birthday );
+  
+  if(!success)  res.status(400).send('Registration failed '+ error);
+  else{res.status(201).send("Registration successful")};
+    
+  }
 
 function createAccessTokenFromRefreshToken(req, res){
     const refreshToken = req.body.refreshToken;
-    if (refreshToken == null) return res.sendStatus(401);
+    if (refreshToken == null) return res.status(400).send('A refresh token is needed');
     jwt.verify(refreshToken,process.env.REFRESH_TOKEN_SECRET, (err, user) => {
-      if (err) return res.sendStatus(403)
+      if (err) return res.status(401).send('The refresh token is invalid')
       const accessToken = generateAccessToken({name : user.name});
-      res.json({accessToken:accessToken});
+      res.status(201).json({accessToken:accessToken});
     }) 
   }
 
@@ -117,12 +59,15 @@ async function logIn(req,res){
     const user = {name : username};
     
     let userAuthenticated = await checkCredentials(username,password);
-    if (!userAuthenticated) return res.sendStatus(401);
-
+    console.log(userAuthenticated)
+    if (!userAuthenticated) {
+      res.status(401).send("Could not find a matching (username, email) - password tuple");
+      return;
+    }
     const accesToken = generateAccessToken(user);
     const refreshToken = generateRefreshToken(user);
     db.addRefreshToken(refreshToken);
-    res.json({accessToken: accesToken, refreshToken: refreshToken})
+    res.status(201).json({accessToken: accesToken, refreshToken: refreshToken})
 }
 
 function logOut(req, res) {
@@ -137,10 +82,10 @@ function logOut(req, res) {
 function authorizeAccessToken(req, res, next) {
     const authHeader = req.headers["authorization"];
     const token = authHeader && authHeader.split(" ")[1];
-    if (token == null) return res.sendStatus(401);
+    if (token == null) return res.sendStatus(400);
   
     jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, user) => {
-      if (err) return  res.sendStatus(403);
+      if (err) return  res.status(401).send('The access token is invalid');
       req.user = user;
       next();
     })
@@ -154,4 +99,4 @@ function generateRefreshToken(user) {
   return jwt.sign(user,process.env.REFRESH_TOKEN_SECRET);
 }
 
-module.exports = {homePage, signup, createAccessTokenFromRefreshToken, logIn, logOut, authorizeAccessToken, googleLogIn}
+module.exports = {homePage, signup, createAccessTokenFromRefreshToken, logIn, logOut, authorizeAccessToken}
